backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
weaviate-client==4.8.0
openai>=1.58.0
python-dotenv==1.0.0

# LangChain ecosystem
langchain>=0.3.0
langchain-text-splitters>=0.3.0
langchain-openai>=0.3.0
langgraph>=0.2.0
langchain-community>=0.3.0

# Observability
langfuse>=2.50.0

# Embeddings
sentence-transformers==3.0.0
torch==2.4.0
transformers>=4.38.0

# Caching & Storage
redis>=5.0.0

# Reranking
flashrank>=0.2.0

# Utilities
pydantic>=2.0.0
